+++ src/registerDynamicSymbols.c
@@ -13,13 +13,11 @@
 
 /* .Call calls */
 extern SEXP _RcppBlaze_blaze_version(SEXP);
-extern SEXP _RcppBlaze_testAs1(SEXP);
-extern SEXP _RcppBlaze_testWrap1();
+extern SEXP _RcppBlaze_fastLmPure(SEXP, SEXP, SEXP);
 
 static const R_CallMethodDef CallEntries[] = {
   {"_RcppBlaze_blaze_version", (DL_FUNC) &_RcppBlaze_blaze_version, 1},
-  {"_RcppBlaze_testAs1", (DL_FUNC) &_RcppBlaze_testAs1, 1},
-  {"_RcppBlaze_testWrap1", (DL_FUNC) &_RcppBlaze_testWrap1, 0},
+  {"_RcppBlaze_fastLmPure", (DL_FUNC) &_RcppBlaze_fastLmPure, 3},
   {NULL, NULL, 0}
 };